requirements.txt
flask
firebase-admin
google-generativeai
python-dotenv